Standardize class names
Archive class names are all over the place.  Standardize the naming convention for these classes and rename the corresponding files wherever appropriate.
